fixed import issues by putting all source files into qclib directory
- this changes how we import qclib. Instead of doing this:

import QC

we need to do this:

from qclib import QC

I think the latter is better, as we namespace all functions to the qclib
package, making it easier to derive where it came from

- not sure how do deal with relative imports within a the package
itself. Now I did "from . import Platforms" and "from .utils.qc_input
import QCInput". Not sure if this is the way to go, or if there are
other alternative solutions