Changed how the overall flag calculation handles extra_flags
It should check for every argument if it's a list or None, Then if it's a list
it should be used in the calculation if it's None it should be ignored.

Also changed the velocity_from_location_list to be stricter. Location now has
to be present, the individual lon and lat can still be None.