Replaced chla local ranges in baltic with global ranges
Commenting out the local range test worked but it set all the final
flags for the data to 0. This is the desired behaviour but the external
interface is filtering out data point with QC=0. Oceanography needs to
properly fix the local ranges but until then we will run with global
ranges.